Environments
Added a link to the solution from the question. This may mess up other chapters when you try to run them - I'm going to update them asap.
How it works:
add a label to the solution: eg. \label{soln:introduction:chemtrails}. In the problem, just add a set of curly brackets before the problem that says {soln:introduction:chemtrails}.

I also changed the solution environment so that when you reference the problem you dont have to do {\ref{prob:...}}, just write {prob:...}.

Trying out a "Reflect and Research" box, really just to add a bit of colour to the "Thinking About the Material" section. May or may not keep it. The environment is called "reflectresearch".