Changes/Updates
Filled in most of the blanks for this chapter. Will be polishing up and adding "Josh's Thoughts" on fields and continuous charges.

There are a few weird errors I am coming across when connecting the solutions to the example questions, and I intend on fixing those by the end of today.