Added ink and removed "fat" images
I found that there were many leftover figured from the accidental copy-paste of the question library folders to the art of modelling folders. I took some time and cleaned those up so that there are no leftover pictures which arent used in each folder. Additionally, the ink files have been slotted in. Was a bit of a headache looking through and making judgement calls about when to put ink files in, but it looks like every drawing which has an associated ink file has the ink file in the folder next to it. Many drawings did not have ink files, but for the ones that had half-completed ink files, I put the "rough" ink file next to the complete drawing. I don't know if you read these, but sorry for it being such a long-winded description, it is just difficult to describe what work was done in a few words.