Circuits done
Finished up circuits. Some formatting issues in "Important Equations" slowed the process down. Also, there weren't any new definitions, so I left that part out of the end of the chapter.

Apologies about this branch not having a name which reflects the work being done in it. I'll be working to sort that out starting with the Magnetic forces chapter and onwards.