Beginning the Wrap Up Process
I decided that instead of doing each chapter individually, I'd go through and write all of the questions, then all of the Josh's thoughts, ect.. So I'm in the process of doing that now. Should be working on figures (which is the last bit of the process) by the end of Wednesday.